+++ psdi_data_conversion/converters/c2x/converter.py
@@ -1,18 +1,18 @@
 """@file psdi_data_conversion/converters/script_template/converter.py
 
-ScriptTemplate file converter
+c2x file converter
 """
 
 from psdi_data_conversion.converters.base import FileConverterMeta, ScriptFileConverter
 
 
-class ScriptTemplateFileConverter(ScriptFileConverter):
-    """File converter specialised to use ScriptTemplate for conversions"""
+class C2xFileConverter(ScriptFileConverter):
+    """File converter specialised to use c2x for conversions"""
 
     meta: FileConverterMeta = FileConverterMeta.load(__file__)
 
-    script = ""
-    required_bin = ""
+    script = "c2x.sh"
+    required_bin = "c2x"
 
     has_in_format_flags_or_options = False
     has_out_format_flags_or_options = False
@@ -21,11 +21,15 @@ class ScriptTemplateFileConverter(ScriptFileConverter):
     allowed_options = ()
 
     def _get_script_args(self):
-        """Optional override: Override the standard script arguments
+        """Override the standard script arguments so we can set the different format names expected by c2x
         """
         l_script_args = super()._get_script_args()
 
-        # Modify the arguments here
+        # Update the output format to c2x style
+        l_script_args[0] = "--" + self.to_format_info.c2x_format
+
+        # TODO - check if the input file has an extension which will be accepted by c2x for its format, and handle if
+        # not
 
         return l_script_args
 
@@ -36,4 +40,4 @@ class ScriptTemplateFileConverter(ScriptFileConverter):
 
 # Assign this converter to the `converter` variable - this lets the psdi_data_conversion.converter module detect and
 # register it, making it available for use by the CLI and web app
-converter = ScriptTemplateFileConverter
+converter = C2xFileConverter