+++ psdi_data_conversion/database.py
@@ -154,7 +154,7 @@ class ConverterInfo:
 
         try:
             self.converter_class = get_registered_converter_class(self.name)
-            self.pretty_name = self.converter_class.name
+            self.pretty_name = self.converter_class.meta.name
         except KeyError:
             self.converter_class = None
             self.pretty_name = name
@@ -182,7 +182,7 @@ class ConverterInfo:
         """Whether or not the converter is ready to be used by this package"""
 
         try:
-            self._key_prefix = get_registered_converter_class(name).database_key_prefix
+            self._key_prefix = get_registered_converter_class(name).meta.database_key_prefix
         except KeyError:
             # We'll get a KeyError for converters in the database that don't yet have their own class, which we can
             # safely ignore
@@ -852,7 +852,7 @@ class ConversionsTable:
 
         # Check if this converter deals with ambiguous formats, so we know if we need to be strict about getting format
         # info
-        if get_registered_converter_class(converter_name).supports_ambiguous_extensions:
+        if get_registered_converter_class(converter_name).meta.supports_ambiguous_extensions:
             which_format = None
         else:
             which_format = 0
@@ -1790,17 +1790,17 @@ def _find_arg(tl_args: tuple[list[FlagInfo], list[OptionInfo]],
                                          "conversion")
 
 
-def get_in_format_args(converter_name: str,
-                       format_name: str,
+def get_in_format_args(converter_name: str | int | UUID | ConverterInfo,
+                       format_name: str | int | UUID | FormatInfo,
                        arg: str | None = None) -> tuple[list[FlagInfo], list[OptionInfo]] | ArgInfo:
     """Get the input flags and options supported by a given converter for a given format (provided as its extension).
     Optionally will provide information on just a single flag or option if its value is provided as an optional argument
 
     Parameters
     ----------
-    converter_name : str
+    converter_name : str | int | UUID | ConverterInfo
         The converter name
-    format_name : str
+    format_name : str | int | UUID | FormatInfo
         The file format name (extension)
     arg : str | None
         If provided, only information on this flag or option will be provided
@@ -1819,18 +1819,18 @@ def get_in_format_args(converter_name: str,
     return _find_arg(tl_args, arg)
 
 
-def get_out_format_args(converter_name: str,
-                        format_name: str,
+def get_out_format_args(converter_name: str | int | UUID | ConverterInfo,
+                        format_name: str | int | UUID | FormatInfo,
                         arg: str | None = None) -> tuple[list[FlagInfo], list[OptionInfo]]:
     """Get the output flags and options supported by a given converter for a given format (provided as its extension).
     Optionally will provide information on just a single flag or option if its value is provided as an optional argument
 
     Parameters
     ----------
-    converter_name : str
-        The converter name
-    format_name : str
-        The file format name (extension)
+    converter_name : str | int | UUID | ConverterInfo
+        The converter
+    format_name : str | int | UUID | FormatInfo
+        The file format (extension)
     arg : str | None
         If provided, only information on this flag or option will be provided