+++ tests/python/converter_test.py
@@ -14,8 +14,8 @@ import pytest
 
 from psdi_data_conversion import constants as const
 from psdi_data_conversion.converter import L_REGISTERED_CONVERTERS
-from psdi_data_conversion.converters.c2x import C2xFileConverter
-from psdi_data_conversion.converters.openbabel import OBFileConverter
+from psdi_data_conversion.converters.c2x.converter import C2xFileConverter
+from psdi_data_conversion.converters.openbabel.converter import OpenBabelFileConverter
 from psdi_data_conversion.database import get_format_info
 from psdi_data_conversion.testing.conversion_test_specs import l_library_test_specs
 from psdi_data_conversion.testing.utils import run_test_conversion_with_library
@@ -71,13 +71,13 @@ def test_envvars():
     assert not math.isclose(converter_no_ev.max_file_size, test_file_size*const.MEGABYTE)
 
     # And check that OB uses its own EV
-    converter = OBFileConverter(filename="1NE6.mmcif",
-                                to_format=pdb_format_id,
-                                use_envvars=True,)
+    converter = OpenBabelFileConverter(filename="1NE6.mmcif",
+                                       to_format=pdb_format_id,
+                                       use_envvars=True,)
     assert not math.isclose(converter.max_file_size, test_file_size*const.MEGABYTE)
 
     os.environ[const.MAX_FILESIZE_OB_EV] = str(test_file_size)
-    converter = OBFileConverter(filename="1NE6.mmcif",
-                                to_format=pdb_format_id,
-                                use_envvars=True,)
+    converter = OpenBabelFileConverter(filename="1NE6.mmcif",
+                                       to_format=pdb_format_id,
+                                       use_envvars=True,)
     assert math.isclose(converter.max_file_size, test_file_size*const.MEGABYTE)