+++ tests/python/database_test.py
@@ -11,9 +11,6 @@ import pytest
 
 from psdi_data_conversion import constants as const
 from psdi_data_conversion.converter import L_SUPPORTED_CONVERTERS
-from psdi_data_conversion.converters.atomsk import CONVERTER_ATO
-from psdi_data_conversion.converters.c2x import CONVERTER_C2X
-from psdi_data_conversion.converters.openbabel import CONVERTER_OB
 from psdi_data_conversion.database import (FileConverterDatabaseException, disambiguate_formats,
                                            get_conversion_pathway, get_conversion_quality, get_converter_info,
                                            get_database, get_format_info, get_in_format_args, get_out_format_args,
@@ -88,7 +85,7 @@ def test_converter_info():
 def test_format_args():
     """Test that we can get the flags and options allowed for specific formats for a given converter
     """
-    converter_name = CONVERTER_OB
+    converter_name = const.CONVERTER_OB
     in_format = "pdb"
     out_format = "cif"
 
@@ -201,17 +198,17 @@ def test_disambiguate_format():
     """
 
     # Test that we can disambiguate the right pdb format
-    in_format, out_format = disambiguate_formats(CONVERTER_OB, "pdb", "cif")
+    in_format, out_format = disambiguate_formats(const.CONVERTER_OB, "pdb", "cif")
     assert in_format == get_format_info("pdb-0")
     assert out_format == get_format_info("cif")
 
     # Test we get the expected exception if no conversion is possible
     with pytest.raises(FileConverterDatabaseException, match="is not supported"):
-        disambiguate_formats(CONVERTER_C2X, "ins", "cml")
+        disambiguate_formats(const.CONVERTER_C2X, "ins", "cml")
 
     # Test we get the expected exception if multiple conversions are possible
     with pytest.raises(FileConverterDatabaseException, match="is ambiguous"):
-        disambiguate_formats(CONVERTER_C2X, "cif", "pdb")
+        disambiguate_formats(const.CONVERTER_C2X, "cif", "pdb")
 
 
 def test_conversion_table():
@@ -225,15 +222,15 @@ def test_conversion_table():
     assert conversions_table.parent is database
 
     # Check we can get the correct conversion quality
-    assert get_conversion_quality(CONVERTER_OB, "pdb", "cif").qual_str == const.QUAL_UNKNOWN
-    assert get_conversion_quality(CONVERTER_ATO, "xyz", "inchi") is None
+    assert get_conversion_quality(const.CONVERTER_OB, "pdb", "cif").qual_str == const.QUAL_UNKNOWN
+    assert get_conversion_quality(const.CONVERTER_ATO, "xyz", "inchi") is None
 
     # Do some detailed checks on one conversion
     xyz_format_info = get_format_info(50135205643343990489495467470022579507)
     inchi_format_info = get_format_info("inchi")
 
     # "xyz" is ambiguous, but only one possibility has a valid conversion here, so check that we get that one
-    qual = get_conversion_quality(CONVERTER_OB, "xyz", inchi_format_info)
+    qual = get_conversion_quality(const.CONVERTER_OB, "xyz", inchi_format_info)
 
     assert qual.qual_str == const.QUAL_OKAY
     assert qual.in_format is get_format_info(xyz_format_info)
@@ -261,11 +258,11 @@ def test_conversion_table():
 
     # Check we can get a list of possible converters for a given conversion
     l_possible_conversions = get_possible_conversions("pdb", "cif")
-    assert (get_converter_info(CONVERTER_OB), get_format_info("pdb", which=0),
+    assert (get_converter_info(const.CONVERTER_OB), get_format_info("pdb", which=0),
             get_format_info("cif", which=0)) in l_possible_conversions
 
     # Check that we can get a list of possible input/outpat formats for a given converter
-    l_in_formats, l_out_formats = get_possible_formats(CONVERTER_OB)
+    l_in_formats, l_out_formats = get_possible_formats(const.CONVERTER_OB)
     assert get_format_info("pdb", which=0) in l_in_formats
     assert get_format_info("cif", which=0) in l_out_formats
 
@@ -284,7 +281,7 @@ def test_conversion_pathways():
     cif_to_inchi_path = get_conversion_pathway("cif", "inchi", only="registered")
     assert len(cif_to_inchi_path) == 1
     converter_info, in_format_info, out_format_info = cif_to_inchi_path[0]
-    assert converter_info.name == regularize_name(CONVERTER_OB)
+    assert converter_info.name == regularize_name(const.CONVERTER_OB)
     assert in_format_info.name == "cif"
     assert out_format_info.name == "inchi"