+++ CHANGELOG.md
@@ -19,7 +19,11 @@
   - Since the graphs of the `ConversionsTable` class (`graph`, `supported_graph`, and `registered_graph`) don't support UUIDs as vertex IDs, they internally use indices for these vertices, which are generated at runtime. To support converting between UUIDs and vertex indices, the following dicts have been added to this class:
     - `d_indices_from_uuids`
     - `d_uuids_from_indices`
+- The converter plugin Python files have been moved from `psdi_data_conversion.converters.{name}` to `psdi_data_conversion.converters.{name}.converter`
 - The `ConversionQualityInfo` object returned by the method `psdi_data_conversion.database.get_conversion_quality` will now have attributes `in_format` and `out_format` always be `FormatInfo` objects, rather than the types of these matching the types of the formats input to this method
+- The internals of the `FileConverter` base class from `psdi_data_conversion.converters.base` have been reworked alongside database changes. This results in most data about an individual converter being stored within its `meta` attribute now, which is of the new `FileConverterMeta` class. The old attributes remain as properties which reference these to avoid breaking existing code as much as possible, but as it's to be deprecated for properties to also be class methods, these will not be accessible from the class itself, and instead must be accessed from the `meta` attribute. That is:
+  - `FileConverter.name` used to work, but now won't. Use `FileConverter.meta.name` instead (or equivalent with any subclass of `FileConverter`)
+  - `x = FileConverter(); x.name` still works. `x.meta.name` can be equivalently be used, but isn't needed
 
 ### New and Changed Functionality
 
@@ -28,7 +32,7 @@
   - Both `False`: The converter is known to exist, but we currently provide no support for it
   - `supported==True`, `registered==False`: This package has a plugin ready to support this converter, but it cannot currently be used due to e.g. a required binary being missing which must be provided by the user
   - Both `True`: The converter is supported and ready to use
-- `ConverterInfo`, `FormatInfo`, `FlagInfo`, and `OptionInfo` now have a `uuid` property which provides the ID converted to a UUID class
+- `ConverterInfo`, `FormatInfo`, `FlagInfo`, `OptionInfo`, and individual converter classes now have a `uuid` property which provides the ID converted to a UUID class
 
 ### Bugfixes